feat: add the runtime loader module
