+++ README.md
@@ -93,6 +93,58 @@ Options:
   -h, --help                        display help for command
 ```
 
+## Runtime Addon Loader
+
+The runtime addon loader allows you to load the addon for the current runtime during runtime.
+
+```ts
+import path from 'path';
+import { fileURLToPath } from 'url';
+import { loadAddon } from 'cmake-ts/dist/loader.mjs';
+
+async function main() {
+  const dirname = typeof __dirname === 'string' ? __dirname : path.dirname(fileURLToPath(import.meta.url));
+
+  const addon = await loadAddon(path.resolve(dirname, '..', 'build'));
+
+  // use the addon
+}
+
+main()
+  .catch((err) => {
+    console.error(err);
+    process.exit(1);
+  });
+```
+
+or in CommonJS:
+
+```js
+const { loadAddon } = require('cmake-ts/dist/loader.js');
+
+async function main() {
+  const addon = await loadAddon(path.resolve(__dirname, '..', 'build'));
+
+  // use the addon
+}
+
+main()
+  .catch((err) => {
+    console.error(err);
+    process.exit(1);
+  });
+```
+
+You can pass the types of the addon to the loader to get type safety:
+
+```ts
+type MyAddon = {
+  myFunction: (name: string) => void;
+};
+
+const addon = await loadAddon<MyAddon>(path.resolve(dirname, '..', 'build'));
+```
+
 ## Configuration File
 
 Configuration is done entirely via `package.json`. You can specify multiple build configurations under the `cmake-ts` key:
