+++ package.json
@@ -60,7 +60,9 @@
     "build.modern-main": "vite build --mode modern-main",
     "build.legacy-library": "vite build --mode legacy-library",
     "build.modern-library": "vite build --mode modern-library",
-    "build": "turbo run build.tsc build.tsc.lib build.legacy-main build.modern-main build.legacy-library build.modern-library",
+    "build.modern-loader": "vite build --mode modern-loader",
+    "build.legacy-loader": "vite build --mode legacy-loader",
+    "build": "turbo run build.tsc build.tsc.lib build.legacy-main build.modern-main build.legacy-library build.modern-library build.modern-loader build.legacy-loader",
     "test.install": "cd test && npm i && cd ..",
     "test": "run-s test.install && cross-env NODE_OPTIONS=--enable-source-maps pnpx vitest --watch false",
     "coverage": "run-s test.install && cross-env NODE_OPTIONS=--enable-source-maps pnpx vitest --coverage --watch false"
