+++ src/build.ts
@@ -1,7 +1,8 @@
 import { join, relative, resolve } from "path"
 import { copy, ensureDir, pathExists, readFile, remove, writeFile } from "fs-extra"
 import { ArgumentBuilder } from "./argumentBuilder.js"
-import { type BuildConfiguration, type Options, getConfigFile, parseBuildConfigs } from "./config.js"
+import type { BuildConfiguration, Options } from "./config-types.d"
+import { getConfigFile, parseBuildConfigs } from "./config.js"
 import { logger } from "./logger.js"
 import { applyOverrides } from "./override.js"
 import { RuntimeDistribution } from "./runtimeDistribution.js"
