+++ src/lib.d.mts
@@ -1,2 +1,2 @@
 export * from "./build.js"
-export * from "./config.js"
+export * from "./config-types.d"
