+++ src/lib.ts
@@ -1,3 +1,4 @@
 export * from "./build.js"
-export * from "./config.js"
+export * from "./config-types.d"
 export * from "./logger.js"
+export * from "./loader.js"
